+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - **[AppLogger]:**: introduce debug instance
+- **[event-viewer]:**: adds debug loging when command execution throws error
 
 ### Fixed
 - **[log-selection]:** event-logs command parsing when thousands separator is '.' instead of ','